+++ framework/src/imgui_utils.cpp
@@ -1,5 +1,5 @@
 #define IMGUI_DEFINE_MATH_OPERATORS
-#include "imgui_user.h"
+#include "imgui_utils.h"
 #include "imgui_internal.h"
 
 void ImGui::ImageWithBg(ImTextureID user_texture_id, const ImVec2& size, const ImVec2& uv0, const ImVec2& uv1, const ImVec4& tint_col, const ImVec4& border_col, const ImVec4& bg_col)
