+++ visual_studio/gears_vk/gears-vk.vcxproj
@@ -46,7 +46,7 @@
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Release_Vulkan|x64'">false</ExcludedFromBuild>
     </ClCompile>
     <ClCompile Include="..\..\framework\src\fixed_update_timer.cpp" />
-    <ClCompile Include="..\..\framework\src\imgui_user.cpp" />
+    <ClCompile Include="..\..\framework\src\imgui_utils.cpp" />
     <ClCompile Include="..\..\framework\src\input_buffer.cpp" />
     <ClCompile Include="..\..\framework\src\log.cpp" />
     <ClCompile Include="..\..\framework\src\material_image_helpers.cpp" />
@@ -156,7 +156,7 @@
     <ClInclude Include="..\..\framework\include\files_changed_event.hpp" />
     <ClInclude Include="..\..\framework\include\gvk.hpp" />
     <ClInclude Include="..\..\framework\include\image_data.hpp" />
-    <ClInclude Include="..\..\framework\include\imgui_user.h" />
+    <ClInclude Include="..\..\framework\include\imgui_utils.h" />
     <ClInclude Include="..\..\framework\include\invokee.hpp" />
     <ClInclude Include="..\..\framework\include\composition.hpp" />
     <ClInclude Include="..\..\framework\include\composition_interface.hpp" />
