+++ visual_studio/gears_vk/gears-vk.vcxproj.filters
@@ -127,7 +127,7 @@
     <ClCompile Include="..\..\external\universal\src\imgui_demo.cpp">
       <Filter>gears-vk_src\external</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\framework\src\imgui_user.cpp">
+    <ClCompile Include="..\..\framework\src\imgui_utils.cpp">
       <Filter>gears-vk_src\user_interface</Filter>
     </ClCompile>
   </ItemGroup>
@@ -519,7 +519,7 @@
     <ClInclude Include="..\..\framework\include\image_data.hpp">
       <Filter>gears-vk_include\data</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\framework\include\imgui_user.h">
+    <ClInclude Include="..\..\framework\include\imgui_utils.h">
       <Filter>gears-vk_include\user_interface</Filter>
     </ClInclude>
   </ItemGroup>
